Add line component extraction and interpolation
Still have some debug conditions to print and visualize stuff, should
remove them in the future.
Parameters are fairly tuned as is, but the extension of the end lines in
order to get the smoothed lines seems to be somewhat struggling. Should
look into some sort of fix for it, in the event that it becomes
problematic.