Add data labeling tools and requirements.txt
Both a tool to label an output during training, as well as a tool to
manually specify the ground truth for a data point is added. Read their
documentation and help messages to figure out how they work.
Added dependency list.
Increased the merge threshold for the lane detection algorithm, so that
it is more robust in real-life images.